refactor: common 구조 수정 및 주석 추가 #17
