+++ src/main/java/com/example/chatserver/common/response/GlobalResponse.java
@@ -2,8 +2,4 @@ package com.example.chatserver.common.response;
 
 import java.time.LocalDateTime;
 
-public record GlobalResponse<T>(
-        boolean success,
-        T data,
-        LocalDateTime timestamp
-) { }
+public record GlobalResponse<T>(boolean success, T data, LocalDateTime timestamp) { }
