+++ src/main/java/com/example/chatserver/common/security/JwtHandshakeHandler.java
@@ -1,5 +1,6 @@
-package com.example.chatserver.common.config;
+package com.example.chatserver.common.security;
 
+import lombok.NonNull;
 import org.springframework.http.server.ServerHttpRequest;
 import org.springframework.stereotype.Component;
 import org.springframework.web.socket.WebSocketHandler;
@@ -9,11 +10,14 @@ import java.security.Principal;
 import java.util.Map;
 
 @Component
-public class HandshakeHandler extends DefaultHandshakeHandler {
+public class JwtHandshakeHandler extends DefaultHandshakeHandler {
 
+    /**
+     * Principal 생성 (name = userId)
+     */
     @Override
-    protected Principal determineUser(ServerHttpRequest request,
-                                      WebSocketHandler wsHandler,
+    protected Principal determineUser(@NonNull ServerHttpRequest request,
+                                      @NonNull WebSocketHandler wsHandler,
                                       Map<String, Object> attributes) {
 
         Object userId = attributes.get(JwtHandshakeInterceptor.ATTR_USER_ID);
